snapp:  rename appbar components
nav menu ref fix
