+++ apps/sensenet/src/components/layout/DesktopLayout.tsx
@@ -4,7 +4,7 @@ import React, { useContext, useEffect, useState } from 'react'
 import { useInjector, useRepository } from '@sensenet/hooks-react'
 import snLogo from '../../assets/sensenet_logo_transparent.png'
 import { ResponsivePersonalSetttings } from '../../context'
-import { DesktopAppBar } from '../appbar/DesktopAppBar'
+import { DesktopAppBar } from '../appbar/desktop-app-bar'
 import { PermanentDrawer } from '../drawer/PermanentDrawer'
 import { TemporaryDrawer } from '../drawer/TemporaryDrawer'
 import { CustomActionCommandProvider } from '../../services/CommandProviders/CustomActionCommandProvider'
